pyproject.toml
[tool.poetry]
name = "nonebot-plugin-ncm"
version = "1.4.0"
description = "基于go-cqhttp与nonebot2的 网易云 无损音乐下载"
license = "Apache License 2.0"
authors = ["kitUIN <kulujun@gmail.com>"]
maintainers = ["kitUIN <kulujun@gmail.com>", "Kurokitu"]
readme = "README.md"
packages = [
    { include = "nonebot-plugin-ncm" }
]
homepage = "https://github.com/kitUIN/nonebot-plugin-ncm"
repository = "https://github.com/kitUIN/nonebot-plugin-ncm"
keywords = ["netease-cloud-music", "netease", "go-cqhttp", "nonebot2"]
[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0b1"
tinydb = "^4.7.0"
pyncm ="^1.6.8.2.2"
aiofile ="^3.7.4"
nonebot-adapter-onebot = "^2.0.0b1"
qrcode = "^7.3.1"
[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
